refactor: main.swift 내 주요 타입 및 기능을 다른 파일로 이동
