+++ contactManager/contactManager/main.swift
@@ -1,64 +1,5 @@
 import Foundation
 
-struct Contact : Hashable {
-    let name : String
-    let age : String
-    let phoneNumber : String
-    
-    var description: String {
-          "- \(self.name) / \(self.age) / \(self.phoneNumber)"
-       }
-}
-
-enum ErrorType{
-    case name
-    case age
-    case phoneNumber
-}
-
-enum GuideText {
-   static let menu = "1) 연락처 추가 2) 연락처 목록보기 3) 연락처 검색 x) 종료"
-   static let selectMenu = "메뉴를 선택해주세요 : "
-}
-
-enum MenuType: String {
-    case addContact = "1"
-    case viewContactList = "2"
-    case findContact  = "3"
-    case exit = "x"
-}
-
-struct Validation {
-    static func validateInputBySlash(input: String) -> Bool {
-        guard (input.filter{ $0 == "/" }.count == 2) else {
-            return false
-        }
-        return true
-    }
-
-    static func validateInputByHyphenNum(input: String) -> Bool {
-        guard (input.filter { $0 == "-" }.count == 2) else {
-            return false
-        }
-        return true
-    }
-}
-
-func errorTextOutput(with type: ErrorType){
-    let errorText: String
-    switch type {
-    case .name: errorText = "이름"
-    case .age: errorText = "나이"
-    case .phoneNumber: errorText = "연락처"
-    }
-    print("입력한 \(errorText)정보가 잘못되었습니다. 입력 형식을 확인해주세요.")
-}
-
-var contacts : Set<Contact> = []
-
-var isMenuInputError: Bool = false
-let regex = "^[A-Za-z\\s]+$"
-
 while !isMenuInputError {
     print(GuideText.menu)
     print(GuideText.selectMenu,terminator: "")
